created styling for news feed and consolidated files
