+++ src/styles/NewsStyles/index.less
@@ -1,0 +1,73 @@
+//---------------------  News Feed styles  --------------------------
+.news-feed-page {
+}
+.news-container {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  border: red solid 1px;
+}
+
+.news-card {
+  width: 80%;
+  height: 250px;
+  margin: 20px 0;
+  padding: 15px;
+  border: black solid 1px;
+  border-radius: 30px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background-color: white;
+  // h1 {
+  //   font-size: 2rem;
+  // }
+
+  // h3 {
+  //   font-size: 1rem;
+  // }
+}
+.title-container {
+  width: 100%;
+  height: 20%;
+  font-size: 1.5rem;
+}
+.description {
+  height: 50%;
+  width: 90%;
+  margin: 1.2% 0;
+}
+.button-container {
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+  height: 20%;
+  width: 90%;
+  a {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-decoration: none;
+    color: black;
+    width: 25%;
+    height: 30px;
+    margin: 0 10%;
+    border: black solid 1px;
+    border-radius: 15px;
+    color: white;
+    background-color: #006c72;
+  }
+  button {
+    text-decoration: none;
+    width: 25%;
+    height: 30px;
+    margin: 0 10%;
+    border: black solid 1px;
+    border-radius: 15px;
+    color: white;
+    background-color: #006c72;
+  }
+}
