+++ src/styles/index.less
@@ -1,6 +1,5 @@
-
-@import "./reset.less";
-@import "./global.less";
-@import "./ParentStyles";
+@import './reset.less';
+@import './global.less';
+@import './ParentStyles';
 @import './InstructorStyles';
-
+@import './NewsStyles/';
